Plots for all animals using goal-actions
As the plots show, every single one of the giraffes and elephants,
sadly, is eaten. The tigers themselves do not survive for too long
because they die mostly from hunger (88%). The source of fish in the
rivers is not great enough to sustain them, and they die out. It's
interesting to see how the elephants do much better when alone (plots
3).